pyproject.toml
[tool.poetry]
name = "ot-harjoitustyo"
version = "0.1.0"
description = ""
authors = ["Eero Ranta <eeroranta111@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pygame = "^2.1.2"
invoke = "^1.7.0"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.1"
coverage = "^6.3.2"
pylint = "^2.13.4"
autopep8 = "^1.6.0"
pytest-dotenv = "^0.5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
